refactor(api): migrate find association repository to its bounded context
